fix typo and var type

+++ nn/src/utils/segmentation.py
@@ -6,5 +6,5 @@ def multi_class_prediction(output: Tensor) -> Tensor:
     return torch.argmax(output, dim=1)
 
 
-def binary_prediction(output: Tensor, thresh: int = 0.0) -> Tensor:
+def binary_prediction(output: Tensor, thresh: float = 0.0) -> Tensor:
     return (output.squeeze(1) > thresh).long()
